# Architecture_DZ_10/API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cloud service for robot-vacuum cleaner",
    "description": "This is a study project considering server based on the OpenAPI 3.0 specification. \nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.",
    "contact": {
      "email": "kardan97@yandex.ru"
    },
    "version": "0.0.2"
  },
  "servers": [
    {
      "url": "https://petstore3.swagger.io/api/v3"
    },
    {
      "url": "https://localhost:8080/api/v1"
    }
  ],
...
  "paths": {
...
    "/room": {
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Add a new room",
        "description": "Add a new room",
        "operationId": "addRoom",
        "requestBody": {
          "description": "Create a new room",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/room/{roomID}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Get room by ID",
        "description": "Get room by ID",
        "operationId": "getRoom",
        "parameters": [
          {
            "name": "roomID",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "description": "Invalid room id supplied"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "put": {
        "tags": [
          "rooms"
        ],
        "summary": "Update room by ID",
        "description": "Update room by ID",
        "operationId": "updateRoom",
        "parameters": [
          {
            "name": "roomID",
            "in": "path",
            "description": "room ID that need to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "description": "Invalid room id supplied"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "delete": {
        "tags": [
          "rooms"
        ],
        "summary": "Delete room by ID",
        "description": "This delete room by ID.",
        "operationId": "deleteRoom",
        "parameters": [
          {
            "name": "roomID",
            "in": "path",
            "description": "The room id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "description": "Invalid room id supplied"
          },
          "404": {
            "description": "Room not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roomname": {
            "type": "string"
          },
          "roomsquare": {
            "type": "string"
          }
        },
        "xml": {
          "name": "room"
        }
      },
...
    },
...
  }
}